# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EVS API",
    "description": "USACE Enterprise Visibility Suite API.\n\nEvery endpoint that replaces an Oracle APEX page carries an `x-apex-page` extension so the\ntraceability matrix (legacy/traceability.csv) can be regenerated from the contract.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/financial/summary": {
      "get": {
        "tags": [
          "financial"
        ],
        "summary": "Financial Summary",
        "operationId": "financial_summary_api_v1_financial_summary_get",
        "parameters": [
          {
            "name": "fiscal_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2026,
              "title": "Fiscal Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "1",
        "x-apex-authorization": "Authenticated User"
      }
    },
    "/api/v1/financial/variance-by-program": {
      "get": {
        "tags": [
          "financial"
        ],
        "summary": "Obligation variance against plan to date, by program",
        "description": "Plan to date is each program's funded amount scaled by the enterprise plan pace at the\nlatest CEFMS period (plan_cumulative at that period over the full-year plan).",
        "operationId": "variance_by_program_api_v1_financial_variance_by_program_get",
        "parameters": [
          {
            "name": "fiscal_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2026,
              "title": "Fiscal Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VarianceByProgram"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "1",
        "x-apex-authorization": "Authenticated User"
      }
    },
...
    "/api/v1/facilities/condition": {
      "get": {
        "tags": [
          "facilities"
        ],
        "summary": "BUILDER component condition rows",
        "description": "Interactive Report parameters as on /projects. Columns: building_id, installation, district, uniformat_section, component_type, ci, bci, deficiency_cost, work_plan_year.",
        "operationId": "condition_api_v1_facilities_condition_get",
        "parameters": [
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "installation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Installation"
            }
          },
          {
            "name": "component_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component Type"
            }
          },
          {
            "name": "max_ci",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only components at or below this CI",
              "title": "Max Ci"
            },
            "description": "Only components at or below this CI"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma separated columns, `-` prefix for descending",
              "title": "Sort"
            },
            "description": "Comma separated columns, `-` prefix for descending"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text search across text columns (IR search bar)",
              "title": "Q"
            },
            "description": "Free text search across text columns (IR search bar)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "description": "`csv` streams the filtered rows",
              "default": "json",
              "title": "Format"
            },
            "description": "`csv` streams the filtered rows"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilitySummary"
                }
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-authorization": "Authenticated User"
      }
    },
...
    "/api/v1/public/locks": {
      "get": {
        "tags": [
          "public-locks"
        ],
        "summary": "List Locks",
        "operationId": "list_locks_api_v1_public_locks_get",
        "parameters": [
          {
            "name": "river_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "River Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-authorization": "none"
      }
    },
    "/api/v1/public/locks/{lock_id}": {
      "get": {
        "tags": [
          "public-locks"
        ],
        "summary": "Get Lock",
        "operationId": "get_lock_api_v1_public_locks__lock_id__get",
        "parameters": [
          {
            "name": "lock_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lock Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-authorization": "none"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppropriationRow": {
        "properties": {
          "appropriation": {
            "type": "string",
            "title": "Appropriation"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "allotted": {
            "type": "number",
            "title": "Allotted"
          },
          "committed": {
            "type": "number",
            "title": "Committed"
          },
          "obligated": {
            "type": "number",
            "title": "Obligated"
          },
          "expended": {
            "type": "number",
            "title": "Expended"
          },
          "expiring_fy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiring Fy"
          }
        },
        "type": "object",
        "required": [
          "appropriation",
          "title",
          "allotted",
          "committed",
          "obligated",
          "expended"
        ],
        "title": "AppropriationRow"
      },
      "AsOf": {
        "properties": {
          "source_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source As Of",
            "description": "Newest timestamp inside the source payload"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At",
            "description": "When the EVS ingestion worker stored it"
          },
          "freshness": {
            "type": "string",
            "enum": [
              "fresh",
              "aging",
              "stale",
              "simulated"
            ],
            "title": "Freshness",
            "default": "fresh"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "live | fixtures | simulated | synthetic",
            "default": "fixtures"
          }
        },
        "type": "object",
        "title": "AsOf",
        "description": "Every dataset carries both the source timestamp and when EVS fetched it."
      },
...
      "ExecutionPoint": {
        "properties": {
          "period": {
            "type": "string",
            "format": "date",
            "title": "Period"
          },
          "plan_cumulative": {
            "type": "number",
            "title": "Plan Cumulative"
          },
          "obligated_cumulative": {
            "type": "number",
            "title": "Obligated Cumulative"
          },
          "expended_cumulative": {
            "type": "number",
            "title": "Expended Cumulative"
          }
        },
        "type": "object",
        "required": [
          "period",
          "plan_cumulative",
          "obligated_cumulative",
          "expended_cumulative"
        ],
        "title": "ExecutionPoint"
      },
      "FacilityRow": {
        "properties": {
          "building_id": {
            "type": "string",
            "title": "Building Id"
          },
          "installation": {
            "type": "string",
            "title": "Installation"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "uniformat_section": {
            "type": "string",
            "title": "Uniformat Section"
          },
          "component_type": {
            "type": "string",
            "title": "Component Type"
          },
          "ci": {
            "type": "number",
            "title": "Ci"
          },
          "bci": {
            "type": "number",
            "title": "Bci"
          },
          "deficiency_cost": {
            "type": "number",
            "title": "Deficiency Cost"
          },
          "work_plan_year": {
            "type": "integer",
            "title": "Work Plan Year"
          }
        },
        "type": "object",
        "required": [
          "building_id",
          "installation",
          "district",
          "uniformat_section",
          "component_type",
          "ci",
          "bci",
          "deficiency_cost",
          "work_plan_year"
        ],
        "title": "FacilityRow"
      },
      "FacilitySummary": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/FacilityRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page"
              },
              {
                "type": "null"
              }
            ]
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "as_of"
        ],
        "title": "FacilitySummary"
      },
...
      "FinancialSummary": {
        "properties": {
          "fiscal_year": {
            "type": "integer",
            "title": "Fiscal Year"
          },
          "execution_curve": {
            "items": {
              "$ref": "#/components/schemas/ExecutionPoint"
            },
            "type": "array",
            "title": "Execution Curve"
          },
          "by_appropriation": {
            "items": {
              "$ref": "#/components/schemas/AppropriationRow"
            },
            "type": "array",
            "title": "By Appropriation"
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "fiscal_year",
          "execution_curve",
          "by_appropriation",
          "as_of"
        ],
        "title": "FinancialSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LockDetail": {
        "properties": {
          "lock_id": {
            "type": "string",
            "title": "Lock Id",
            "description": "riverCode + '-' + lockNo, e.g. OH-79"
          },
          "river_code": {
            "type": "string",
            "title": "River Code"
          },
          "river_name": {
            "type": "string",
            "title": "River Name"
          },
          "lock_name": {
            "type": "string",
            "title": "Lock Name"
          },
          "lock_no": {
            "type": "string",
            "title": "Lock No"
          },
          "river_mile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "River Mile"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "chambers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chambers"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "status": {
            "type": "string",
            "enum": [
              "operating",
              "delayed",
              "closed",
              "stale",
              "unknown"
            ],
            "title": "Status"
          },
          "status_reason": {
            "type": "string",
            "title": "Status Reason"
          },
          "vessels_queued": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessels Queued"
          },
          "avg_delay_4h_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Delay 4H Min"
          },
          "avg_delay_24h_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Delay 24H Min"
          },
          "last_lockage_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Lockage At"
          },
          "gauge_stage_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gauge Stage Ft"
          },
          "flood_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flood Category"
          },
          "active_stoppages": {
            "type": "integer",
            "title": "Active Stoppages",
            "default": 0
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          },
          "lift_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lift Ft"
          },
          "chamber_dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chamber Dimensions"
          },
          "year_opened": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Opened"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "queue": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Queue",
            "default": []
          },
          "stoppages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Stoppages",
            "default": []
          },
          "recent_lockages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Lockages",
            "default": []
          },
          "status_inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Status Inputs",
            "description": "inputs_used by the status engine"
          }
        },
        "type": "object",
        "required": [
          "lock_id",
          "river_code",
          "river_name",
          "lock_name",
          "lock_no",
          "river_mile",
          "district",
          "chambers",
          "latitude",
          "longitude",
          "status",
          "status_reason",
          "vessels_queued",
          "avg_delay_4h_min",
          "avg_delay_24h_min",
          "last_lockage_at",
          "as_of"
        ],
        "title": "LockDetail"
      },
      "LockList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LockSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "rivers": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Rivers"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "operating",
                "delayed",
                "closed",
                "stale",
                "unknown"
              ]
            },
            "type": "object",
            "title": "Counts"
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "items",
          "rivers",
          "counts",
          "as_of"
        ],
        "title": "LockList"
      },
...
      "Page": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "title": "Page"
      },
...
      "ProgramVariance": {
        "properties": {
          "program_code": {
            "type": "string",
            "title": "Program Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "business_line": {
            "type": "string",
            "title": "Business Line"
          },
          "division": {
            "type": "string",
            "title": "Division"
          },
          "funded_amount": {
            "type": "number",
            "title": "Funded Amount"
          },
          "obligated_amount": {
            "type": "number",
            "title": "Obligated Amount"
          },
          "plan_to_date": {
            "type": "number",
            "title": "Plan To Date"
          },
          "variance_amount": {
            "type": "number",
            "title": "Variance Amount"
          },
          "variance_pct": {
            "type": "number",
            "title": "Variance Pct"
          }
        },
        "type": "object",
        "required": [
          "program_code",
          "name",
          "business_line",
          "division",
          "funded_amount",
          "obligated_amount",
          "plan_to_date",
          "variance_amount",
          "variance_pct"
        ],
        "title": "ProgramVariance"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VarianceByProgram": {
        "properties": {
          "fiscal_year": {
            "type": "integer",
            "title": "Fiscal Year"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ProgramVariance"
            },
            "type": "array",
            "title": "Rows"
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "fiscal_year",
          "rows",
          "as_of"
        ],
        "title": "VarianceByProgram",
        "description": "Diverging variance bars on `/programs` and `/financial`."
      }
    }
  },
...
}